Add cache compression
It doesn't do a lot in normal use given that caches are usually a few hundred bytes, but with 102 entries the uncompressed cache was 3.6kb and the compressed one was 1.1kb so it does matter more at high cache sizes. Again, not much benefit in regular use cases but it does have some benefits and gives me something to add to the next changelog.