Quick tweak to make the caching more reliable
Essentially, instead of just writing the cache when onSaveInstanceState(...) is called, which was not reliable since a crash or other ungraceful stop would not trigger a cache write, the cache is now written whenever the contents change (i.e. after relapses, deletions and additions). This means that if you log a relapse and then crash the app somehow the relapse will have been cached.